pyproject.toml
[tool.poetry]
name = "super-drf"
version = "0.1.0"
description = "My super duper django rest backend for myself :>"
authors = ["minlaxz <minminlaxz@gmail.com>"]
license = "UNLICENSE"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.4"
django-fsm = "^2.8.0"
graphviz = "^0.20"
django-cors-headers = "^3.11.0"
psycopg2-binary = "^2.9.3"
gunicorn = "^20.1.0"
whitenoise = "^6.2.0"
dj-database-url = "^0.5.0"
django-prometheus = "^2.2.0"
django-debug-toolbar = "^3.7.0"
django-rest-knox = "^4.2.0"
djangorestframework = "^3.14.0"
dj-rest-auth = "^2.2.5"
django-allauth = "^0.51.0"
djangorestframework-simplejwt = "^5.2.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
